добавил контекстные действия со словами
