+++ MadDogsInDaPlace/app/src/main/res/layout/activity_main.xml
@@ -47,4 +47,16 @@
         android:background="#7d010000"
         android:layout_marginBottom="5dp"
         android:onClick="onStartButtonClick" />
+
+    <EditText
+        android:id="@+id/detectedText"
+        style="@style/Widget.AppCompat.EditText"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_above="@+id/startButton"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentStart="true"
+        android:background="#FFFFFFFF"
+        android:inputType="text"
+        android:textSize="30sp" />
 </RelativeLayout>
