+++ MadDogsInDaPlace/app/src/main/res/menu/action_mode_layout.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:onClick="onContextualMenuItemClicked"
+        android:id="@+id/cab_add"
+        android:icon="@android:drawable/ic_menu_add"
+        android:orderInCategory="1"
+        app:showAsAction="ifRoom"
+        android:title="add"/>
+    <item
+        android:onClick="onContextualMenuItemClicked"
+        android:id="@+id/cab_share"
+        android:icon="@android:drawable/ic_menu_share"
+        android:orderInCategory="1"
+        app:showAsAction="ifRoom"
+        android:title="share"/>
+</menu>
